+++ src/main/scala/com/codahale/jerkson/deser/CaseClassDeserializer.scala
@@ -27,7 +27,8 @@ class CaseClassDeserializer(config: DeserializationConfig,
       jp.nextToken()
     }
 
-    if (jp.getCurrentToken != JsonToken.FIELD_NAME) {
+    if (jp.getCurrentToken != JsonToken.FIELD_NAME &&
+      jp.getCurrentToken != JsonToken.END_OBJECT) {
       throw ctxt.mappingException(javaType.getRawClass)
     }