+++ src/main/scala/com/codahale/jerkson/ser/IterableSerializer.scala
@@ -8,17 +8,16 @@ import org.codehaus.jackson.map.{SerializerProvider, JsonSerializer}
  * @author coda
  */
 class IterableSerializer extends JsonSerializer[Iterable[_]] {
-  def serialize(value: Iterable[_], json: JsonGenerator, provider: SerializerProvider) = {
+  def serialize(value: Iterable[_], json: JsonGenerator, provider: SerializerProvider) {
     json.writeStartArray()
     for (element <- value) {
       if (element == null) {
         provider.getNullValueSerializer.serialize(null, json, provider)
       } else {
         val obj = element.asInstanceOf[Object]
-        val serializer = provider.findValueSerializer(obj.getClass)
+        val serializer = provider.findValueSerializer(obj.getClass, null)
         serializer.serialize(obj, json, provider)
       }
-      json.flush()
     }
     json.writeEndArray()
   }