+++ src/main/scala/com/codahale/jerkson/ser/IteratorSerializer.scala
@@ -5,17 +5,16 @@ import org.codehaus.jackson.map.{SerializerProvider, JsonSerializer}
 
 class IteratorSerializer extends JsonSerializer[Iterator[_]] {
   def serialize(value: Iterator[_], json: JsonGenerator,
-                provider: SerializerProvider) = {
+                provider: SerializerProvider) {
     json.writeStartArray()
     for (element <- value) {
       if (element == null) {
         provider.getNullValueSerializer.serialize(null, json, provider)
       } else {
         val obj = element.asInstanceOf[Object]
-        val serializer = provider.findValueSerializer(obj.getClass)
+        val serializer = provider.findValueSerializer(obj.getClass, null)
         serializer.serialize(obj, json, provider)
       }
-      json.flush()
     }
     json.writeEndArray()
   }