+++ src/main/scala/com/codahale/jerkson/ser/JValueSerializer.scala
@@ -6,9 +6,9 @@ import com.codahale.jerkson.AST._
 import java.math.BigInteger
 
 class JValueSerializer extends JsonSerializer[JValue] {
-  def serialize(value: JValue, json: JsonGenerator, provider: SerializerProvider) = {
+  def serialize(value: JValue, json: JsonGenerator, provider: SerializerProvider) {
     value match {
-      case JInt(v) => json.writeNumber(new BigInteger(v.toString))
+      case JInt(v) => json.writeNumber(new BigInteger(v.toString()))
       case JFloat(v) => json.writeNumber(v)
       case JString(v) => json.writeString(v)
       case JBoolean(v) => json.writeBoolean(v)